pyproject.toml
[project]
name = "plataforma-logistica"
version = "0.1.0"
description = "API REST para gestionar envíos terrestres y marítimos"
requires-python = ">=3.11"

dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.0",
  "SQLAlchemy>=2.0.0",
  "psycopg[binary]>=3.2.0",
  "PyJWT>=2.9.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "httpx>=0.27.0",
  "ruff>=0.6.0",
]

[tool.ruff]
line-length = 100

target-version = "py311"

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["."]
